requirements.txt
pandas
numpy
minio
pyspark==3.5.0
python-dotenv==0.21.0
requests==2.26.0